Fixed some warnings

+++ clfsm/machines/RoboCup2013/SMBallFollower.machine/SMBallFollower_Variables.h
@@ -13,4 +13,4 @@ int	boundaryCloseFar;	///<
 SENSORS_TorsoJointSensors_t	body_sensors_ptr;	///< 
 SENSORS_TorsoJointSensors	body_sensors;	///< 
 WALK_Command_t	walk_post;	///< 
-int	limiter;	///<
+float	limiter;	///<
