+++ clfsm/machines/RoboCup2013/SMBallFollower.machine/State_Calc_BallAngle_OnEntry.mm
@@ -1,3 +1,3 @@
 ball = sightings_ptr.get();
 
-ballAngle = RAD2DEG(ball.horizontal_angle());
+ballAngle = (float)RAD2DEG(ball.horizontal_angle());
